+++ src/ui/hitsur.cpp
@@ -126,9 +126,9 @@ void HitSur::Ellipse2(const float centerx, const float centery, const int radius
    const int ix = SCALEXf(centerx);
    const int iy = SCALEYf(centery);
 
-   const long long dx = m_hitx - ix;
-   const long long dy = m_hity - iy;
-   const long long dist = dx*dx + dy*dy;
+   const int dx = m_hitx - ix;
+   const int dy = m_hity - iy;
+   const int dist = dx*dx + dy*dy;
 
    if (dist <= radius*radius)
    {
